fix: remove all consoles and rename file to follow camel case convention
